// Frontend/src/app/models/address.model.ts
export type Address = {
  _id?: string;
  fullName: string;
  phoneNo: string;
  userId?: string;
  country: string;
  state: string;
  city: string;
  addressLane1: string;
  addressLane2?: string;
  landmark?: string;
  pincode: string;
  createdAt?: string;
  updatedAt?: string;
  __v?: string;
  isDefault?: boolean;
};